+++ crates/compiler/src/bytecode_assembler.rs
@@ -607,6 +607,7 @@ pub fn parse_asm_line(bc: &mut Bytecode, line: &str) -> Result<(), ChannelError>
        adjust_fir_tap 2 uvnt_and_svnt_arguments,
        adjust_fir_tap_limit 3 adjust_fir_tap_limit_arguments,
        set_echo_invert 1 parse_invert_flag_arguments,
+       set_echo_delay 1 one_uvnt_argument,
     )
 }