+++ crates/compiler/src/mml/tokenizer.rs
@@ -97,6 +97,7 @@ pub enum Token<'a> {
     FtapPlus,
     FtapMinus,
     SetEchoInvert(InvertFlags),
+    SetEchoDelay,
 
     // Temp GAIN after quantization tokens
     // (GainModeE is the Echo token)
@@ -542,6 +543,7 @@ fn next_token<'a>(scanner: &mut Scanner<'a>) -> Option<TokenWithPosition<'a>> {
                 "ftap" => Token::Ftap,
                 "ftap+" => Token::FtapPlus,
                 "ftap-" => Token::FtapMinus,
+                "edl" => Token::SetEchoDelay,
                 "ei" => {
                     scanner.skip_whitespace();