+++ crates/compiler/tests/mml/bc_asm.rs
@@ -2,6 +2,8 @@
 //
 // SPDX-License-Identifier: MIT
 
+use compiler::echo::EchoEdl;
+
 use crate::*;
 
 #[test]
@@ -240,7 +242,10 @@ ADEF \asm { set_instrument dummy_instrument | play_note a4 24 }
         &dummy_data.instruments_and_samples,
         mml.subroutines(),
         BcTerminator::DisableChannel,
-        BytecodeContext::SongChannel { index: 0 },
+        BytecodeContext::SongChannel {
+            index: 0,
+            max_edl: EchoEdl::MIN,
+        },
     )
     .repeat(4);