+++ crates/compiler/tests/mml/echo.rs
@@ -671,3 +671,42 @@ fn set_echo_invert() {
     assert_one_error_in_mml_line(r"\ei", 1, ValueError::NoInvertFlags.into());
     assert_one_error_in_mml_line(r"\ei c", 1, ValueError::InvalidMmlInvertFlags.into());
 }
+
+#[test]
+fn set_echo_delay() {
+    assert_line_matches_bytecode(r"\edl 0", &["set_echo_delay 0"]);
+
+    assert_mml_channel_a_matches_bytecode_max_edl(
+        r#"
+#MaxEchoLength 48
+
+A \edl 48
+"#,
+        &["set_echo_delay 48"],
+        EchoEdl::try_from(3u8).unwrap(),
+    );
+
+    assert_mml_channel_a_matches_bytecode_max_edl(
+        r#"
+#MaxEchoLength 240
+
+A \edl 240
+"#,
+        &["set_echo_delay 240"],
+        EchoEdl::try_from(15u8).unwrap(),
+    );
+
+    assert_one_error_in_channel_a_mml(
+        r#"
+#MaxEchoLength 16
+
+A \edl 32
+"#,
+        3,
+        BytecodeError::EchoLengthLargerThanMaxEdl {
+            edl: 2u8.try_into().unwrap(),
+            max_edl: 1u8.try_into().unwrap(),
+        }
+        .into(),
+    );
+}