+++ crates/compiler/tests/mml/parsing.rs
@@ -58,7 +58,10 @@ ADEF @0 a
         &dummy_data.instruments_and_samples,
         mml.subroutines(),
         BcTerminator::DisableChannel,
-        BytecodeContext::SongChannel { index: 0 },
+        BytecodeContext::SongChannel {
+            index: 0,
+            max_edl: EchoEdl::MIN,
+        },
     )
     .repeat(4);